+++ src/Facade/AbstractFacade.php
@@ -40,8 +40,9 @@ abstract class AbstractFacade
     /**
      * Create a new mock builder.
      *
-     * The `$types` argument may be a class name, a reflection class, or a mock
-     * builder. It may also be an array of any of these.
+     * The `$types` argument may be an object, a class name, a reflection class,
+     * an anonymous class, or a mock builder. It may also be an array of any of
+     * these.
      *
      * If `$types` is omitted, or `null`, no existing type will be used when
      * generating the mock class. This is useful in the case of ad hoc mocks,
@@ -67,8 +68,9 @@ abstract class AbstractFacade
     /**
      * Create a new full mock.
      *
-     * The `$types` argument may be a class name, a reflection class, or a mock
-     * builder. It may also be an array of any of these.
+     * The `$types` argument may be an object, a class name, a reflection class,
+     * an anonymous class, or a mock builder. It may also be an array of any of
+     * these.
      *
      * If `$types` is omitted, or `null`, no existing type will be used when
      * generating the mock class. This is useful in the case of ad hoc mocks,
@@ -96,8 +98,9 @@ abstract class AbstractFacade
     /**
      * Create a new partial mock.
      *
-     * The `$types` argument may be a class name, a reflection class, or a mock
-     * builder. It may also be an array of any of these.
+     * The `$types` argument may be an object, a class name, a reflection class,
+     * an anonymous class, or a mock builder. It may also be an array of any of
+     * these.
      *
      * If `$types` is omitted, or `null`, no existing type will be used when
      * generating the mock class. This is useful in the case of ad hoc mocks,