+++ src/Mock/Builder/Factory/MockBuilderFactory.php
@@ -84,8 +84,9 @@ class MockBuilderFactory implements MockBuilderFactoryInterface
     /**
      * Create a new mock builder.
      *
-     * The `$types` argument may be a class name, a reflection class, or a mock
-     * builder. It may also be an array of any of these.
+     * The `$types` argument may be an object, a class name, a reflection class,
+     * an anonymous class, or a mock builder. It may also be an array of any of
+     * these.
      *
      * If `$types` is omitted, or `null`, no existing type will be used when
      * generating the mock class. This is useful in the case of ad hoc mocks,
@@ -114,8 +115,9 @@ class MockBuilderFactory implements MockBuilderFactoryInterface
     /**
      * Create a new full mock.
      *
-     * The `$types` argument may be a class name, a reflection class, or a mock
-     * builder. It may also be an array of any of these.
+     * The `$types` argument may be an object, a class name, a reflection class,
+     * an anonymous class, or a mock builder. It may also be an array of any of
+     * these.
      *
      * If `$types` is omitted, or `null`, no existing type will be used when
      * generating the mock class. This is useful in the case of ad hoc mocks,
@@ -138,8 +140,9 @@ class MockBuilderFactory implements MockBuilderFactoryInterface
     /**
      * Create a new partial mock.
      *
-     * The `$types` argument may be a class name, a reflection class, or a mock
-     * builder. It may also be an array of any of these.
+     * The `$types` argument may be an object, a class name, a reflection class,
+     * an anonymous class, or a mock builder. It may also be an array of any of
+     * these.
      *
      * If `$types` is omitted, or `null`, no existing type will be used when
      * generating the mock class. This is useful in the case of ad hoc mocks,