+++ src/Mock/Builder/Factory/MockBuilderFactoryInterface.php
@@ -23,8 +23,9 @@ interface MockBuilderFactoryInterface
     /**
      * Create a new mock builder.
      *
-     * The `$types` argument may be a class name, a reflection class, or a mock
-     * builder. It may also be an array of any of these.
+     * The `$types` argument may be an object, a class name, a reflection class,
+     * an anonymous class, or a mock builder. It may also be an array of any of
+     * these.
      *
      * If `$types` is omitted, or `null`, no existing type will be used when
      * generating the mock class. This is useful in the case of ad hoc mocks,
@@ -45,8 +46,9 @@ interface MockBuilderFactoryInterface
     /**
      * Create a new full mock.
      *
-     * The `$types` argument may be a class name, a reflection class, or a mock
-     * builder. It may also be an array of any of these.
+     * The `$types` argument may be an object, a class name, a reflection class,
+     * an anonymous class, or a mock builder. It may also be an array of any of
+     * these.
      *
      * If `$types` is omitted, or `null`, no existing type will be used when
      * generating the mock class. This is useful in the case of ad hoc mocks,
@@ -67,8 +69,9 @@ interface MockBuilderFactoryInterface
     /**
      * Create a new partial mock.
      *
-     * The `$types` argument may be a class name, a reflection class, or a mock
-     * builder. It may also be an array of any of these.
+     * The `$types` argument may be an object, a class name, a reflection class,
+     * an anonymous class, or a mock builder. It may also be an array of any of
+     * these.
      *
      * If `$types` is omitted, or `null`, no existing type will be used when
      * generating the mock class. This is useful in the case of ad hoc mocks,